changes: shadcn to use current utils and moved data.ts to utils
