+++ frontend/src/components/Card.tsx
@@ -1,12 +1,12 @@
 import { FontAwesomeIcon, FontAwesomeIconProps } from '@fortawesome/react-fontawesome'
 import { useState, useEffect } from 'react'
 import { Tooltip } from 'react-tooltip'
+import { Icons } from 'utils/data'
 import { CardProps, tooltipStyle } from 'lib/constants'
 import FontAwesomeIconWrapper from 'lib/FontAwesomeIconWrapper'
 import { cn } from 'lib/utils'
 import ActionButton from 'components/ActionButton'
 import ContributorAvatar from 'components/ContributorAvatar'
-import { Icons } from 'components/data'
 import DisplayIcon from 'components/DisplayIcon'
 import Markdown from 'components/MarkdownWrapper'
 import TopicBadge from 'components/TopicBadge'
