+++ frontend/src/pages/ProjectDetails.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
+import { level } from 'utils/data'
 import { fetchAlgoliaData } from 'lib/api'
 import FontAwesomeIconWrapper from 'lib/FontAwesomeIconWrapper'
 import { getFilteredIcons } from 'lib/utils'
 import Card from 'components/Card'
-import { level } from 'components/data'
 import LoadingSpinner from 'components/LoadingSpinner'
 
 const ProjectDetailsPage = () => {
