+++ frontend/src/pages/Projects.tsx
@@ -1,10 +1,10 @@
 import { useSearchPage } from 'hooks/useSearchPage'
 import { useNavigate } from 'react-router-dom'
 import { project } from 'types/project'
+import { level } from 'utils/data'
 import FontAwesomeIconWrapper from 'lib/FontAwesomeIconWrapper'
 import { getFilteredIcons } from 'lib/utils'
 import Card from 'components/Card'
-import { level } from 'components/data'
 import SearchPageLayout from 'components/SearchPageLayout'
 
 const ProjectsPage = () => {
